codetrek_backend/requirements.txt
chromadb==0.6.3
Django==5.1.7
djangorestframework==3.15.2
fuzzywuzzy==0.18.0
pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3

